refactor core server/client/facilitator
