+++ python/x402/client.py
@@ -7,345 +7,45 @@ Async is the default with full async hook support.
 from __future__ import annotations
 
 import asyncio
-from collections.abc import Awaitable, Callable
-from dataclasses import dataclass, field
 from typing import Any
 
 from typing_extensions import Self
 
-from .interfaces import SchemeNetworkClient, SchemeNetworkClientV1
+# Re-export from client_base for external use
+from .client_base import (
+    AfterPaymentCreationHook,
+    BeforePaymentCreationHook,
+    OnPaymentCreationFailureHook,
+    PaymentRequirementsSelector,
+    SchemeRegistration,
+    SyncAfterPaymentCreationHook,
+    SyncBeforePaymentCreationHook,
+    SyncOnPaymentCreationFailureHook,
+    default_payment_selector,
+    max_amount,
+    prefer_network,
+    prefer_scheme,
+    x402ClientBase,
+    x402ClientConfig,
+)
 from .schemas import (
-    AbortResult,
-    Network,
-    NoMatchingRequirementsError,
-    PaymentAbortedError,
-    PaymentCreatedContext,
-    PaymentCreationContext,
-    PaymentCreationFailureContext,
     PaymentPayload,
     PaymentPayloadV1,
     PaymentRequired,
     PaymentRequiredV1,
-    PaymentRequirements,
-    PaymentRequirementsV1,
-    RecoveredPayloadResult,
     ResourceInfo,
-    SchemeNotFoundError,
-    find_schemes_by_network,
 )
 
-# ============================================================================
-# Type Aliases
-# ============================================================================
-
-# V2 types
-Requirements = PaymentRequirements
-RequirementsView = PaymentRequirements | PaymentRequirementsV1
-
-# Policy: filter requirements list (e.g., prefer_network, max_amount)
-PaymentPolicy = Callable[[int, list[RequirementsView]], list[RequirementsView]]
-
-# Selector: choose final requirement from filtered list
-PaymentRequirementsSelector = Callable[[int, list[RequirementsView]], RequirementsView]
-
-
-# ============================================================================
-# Configuration Types (matching TypeScript)
-# ============================================================================
-
-
-@dataclass
-class SchemeRegistration:
-    """Configuration for registering a payment scheme with a specific network.
-
-    Matches TypeScript SchemeRegistration interface.
-
-    Attributes:
-        network: The network identifier (e.g., 'eip155:8453', 'solana:mainnet').
-        client: The scheme client implementation for this network.
-        x402_version: The x402 protocol version to use for this scheme.
-            Defaults to 2.
-    """
-
-    network: Network
-    client: SchemeNetworkClient | SchemeNetworkClientV1
-    x402_version: int = 2
-
-
-@dataclass
-class x402ClientConfig:
-    """Configuration options for creating x402Client from config.
-
-    Matches TypeScript x402ClientConfig interface.
-
-    Attributes:
-        schemes: Array of scheme registrations defining which payment methods
-            are supported.
-        policies: Optional policies to apply to the client.
-        payment_requirements_selector: Optional custom payment requirements
-            selector function. If not provided, uses the default selector
-            (first available option).
-    """
-
-    schemes: list[SchemeRegistration]
-    policies: list[PaymentPolicy] | None = None
-    payment_requirements_selector: PaymentRequirementsSelector | None = field(default=None)
-
-
-# Hook types - support both sync and async (for async class auto-detection)
-BeforePaymentCreationHook = Callable[
-    [PaymentCreationContext], Awaitable[AbortResult | None] | AbortResult | None
+__all__ = [
+    "x402Client",
+    "x402ClientSync",
+    "x402ClientConfig",
+    "SchemeRegistration",
+    "default_payment_selector",
+    "prefer_network",
+    "prefer_scheme",
+    "max_amount",
 ]
-AfterPaymentCreationHook = Callable[[PaymentCreatedContext], Awaitable[None] | None]
-OnPaymentCreationFailureHook = Callable[
-    [PaymentCreationFailureContext],
-    Awaitable[RecoveredPayloadResult | None] | RecoveredPayloadResult | None,
-]
-
-# Sync-only hook types (for sync class)
-SyncBeforePaymentCreationHook = Callable[[PaymentCreationContext], AbortResult | None]
-SyncAfterPaymentCreationHook = Callable[[PaymentCreatedContext], None]
-SyncOnPaymentCreationFailureHook = Callable[
-    [PaymentCreationFailureContext], RecoveredPayloadResult | None
-]
-
-
-# ============================================================================
-# Default Implementations
-# ============================================================================
-
-
-def default_payment_selector(
-    version: int,
-    requirements: list[RequirementsView],
-) -> RequirementsView:
-    """Default selector: return first requirement.
-
-    Args:
-        version: Protocol version.
-        requirements: List of filtered requirements.
-
-    Returns:
-        First requirement in list.
-    """
-    return requirements[0]
-
-
-# ============================================================================
-# Built-in Policies
-# ============================================================================
-
-
-def prefer_network(network: Network) -> PaymentPolicy:
-    """Create policy that prefers a specific network.
-
-    Args:
-        network: Network to prefer.
-
-    Returns:
-        Policy function that moves matching requirements to front.
-    """
-
-    def policy(version: int, reqs: list[RequirementsView]) -> list[RequirementsView]:
-        preferred = [r for r in reqs if r.network == network]
-        others = [r for r in reqs if r.network != network]
-        return preferred + others
-
-    return policy
-
-
-def prefer_scheme(scheme: str) -> PaymentPolicy:
-    """Create policy that prefers a specific scheme.
-
-    Args:
-        scheme: Scheme to prefer.
-
-    Returns:
-        Policy function that moves matching requirements to front.
-    """
-
-    def policy(version: int, reqs: list[RequirementsView]) -> list[RequirementsView]:
-        preferred = [r for r in reqs if r.scheme == scheme]
-        others = [r for r in reqs if r.scheme != scheme]
-        return preferred + others
-
-    return policy
-
-
-def max_amount(max_value: int) -> PaymentPolicy:
-    """Create policy that filters by maximum amount.
-
-    Args:
-        max_value: Maximum amount in smallest unit.
-
-    Returns:
-        Policy function that removes requirements exceeding max.
-    """
-
-    def policy(version: int, reqs: list[RequirementsView]) -> list[RequirementsView]:
-        return [r for r in reqs if int(r.get_amount()) <= max_value]
-
-    return policy
-
-
-# ============================================================================
-# Base Client Class (Shared Logic)
-# ============================================================================
-
-
-class _x402ClientBase:
-    """Base class with shared logic for x402 clients.
-
-    Contains registration, policies, and selection logic.
-    Subclasses implement sync/async payment creation.
-    """
-
-    def __init__(
-        self,
-        payment_requirements_selector: PaymentRequirementsSelector | None = None,
-    ) -> None:
-        """Initialize base client.
-
-        Args:
-            payment_requirements_selector: Custom selector for choosing
-                from filtered requirements. Defaults to first match.
-        """
-        self._selector = payment_requirements_selector or default_payment_selector
-        self._schemes: dict[Network, dict[str, SchemeNetworkClient]] = {}
-        self._schemes_v1: dict[Network, dict[str, SchemeNetworkClientV1]] = {}
-        self._policies: list[PaymentPolicy] = []
-
-        # Hooks (typed in subclasses)
-        self._before_payment_creation_hooks: list[Any] = []
-        self._after_payment_creation_hooks: list[Any] = []
-        self._on_payment_creation_failure_hooks: list[Any] = []
-
-    # ========================================================================
-    # Registration
-    # ========================================================================
-
-    def register(self, network: Network, client: SchemeNetworkClient) -> Self:
-        """Register a V2 scheme client for a network.
-
-        Args:
-            network: Network to register for (e.g., "eip155:8453" or "eip155:*").
-            client: Scheme client implementation.
-
-        Returns:
-            Self for chaining.
-        """
-        if network not in self._schemes:
-            self._schemes[network] = {}
-        self._schemes[network][client.scheme] = client
-        return self
-
-    def register_v1(self, network: Network, client: SchemeNetworkClientV1) -> Self:
-        """Register a V1 scheme client for a network.
-
-        Args:
-            network: Network to register for.
-            client: V1 scheme client implementation.
-
-        Returns:
-            Self for chaining.
-        """
-        if network not in self._schemes_v1:
-            self._schemes_v1[network] = {}
-        self._schemes_v1[network][client.scheme] = client
-        return self
-
-    def register_policy(self, policy: PaymentPolicy) -> Self:
-        """Add a requirement filter policy.
-
-        Policies are applied in registration order to filter and reorder
-        payment requirements before selection.
-
-        Args:
-            policy: Policy function to add.
-
-        Returns:
-            Self for chaining.
-        """
-        self._policies.append(policy)
-        return self
-
-    # ========================================================================
-    # Selection (Shared)
-    # ========================================================================
-
-    def _select_requirements_v2(
-        self,
-        requirements: list[PaymentRequirements],
-    ) -> PaymentRequirements:
-        """Select V2 requirements using policies and selector."""
-        # Filter to supported schemes
-        supported = []
-        for req in requirements:
-            schemes = find_schemes_by_network(self._schemes, req.network)
-            if schemes and req.scheme in schemes:
-                supported.append(req)
-
-        if not supported:
-            raise NoMatchingRequirementsError("No payment requirements match registered schemes")
-
-        # Apply policies
-        filtered: list[RequirementsView] = list(supported)
-        for policy in self._policies:
-            filtered = policy(2, filtered)
-            if not filtered:
-                raise NoMatchingRequirementsError("All requirements filtered out by policies")
-
-        # Select final
-        return self._selector(2, filtered)  # type: ignore[return-value]
-
-    def _select_requirements_v1(
-        self,
-        requirements: list[PaymentRequirementsV1],
-    ) -> PaymentRequirementsV1:
-        """Select V1 requirements using policies and selector."""
-        # Filter to supported schemes
-        supported = []
-        for req in requirements:
-            schemes = find_schemes_by_network(self._schemes_v1, req.network)
-            if schemes and req.scheme in schemes:
-                supported.append(req)
-
-        if not supported:
-            raise NoMatchingRequirementsError("No payment requirements match registered schemes")
-
-        # Apply policies
-        filtered: list[RequirementsView] = list(supported)
-        for policy in self._policies:
-            filtered = policy(1, filtered)
-            if not filtered:
-                raise NoMatchingRequirementsError("All requirements filtered out by policies")
-
-        # Select final
-        return self._selector(1, filtered)  # type: ignore[return-value]
-
-    # ========================================================================
-    # Introspection
-    # ========================================================================
-
-    def get_registered_schemes(
-        self,
-    ) -> dict[int, list[dict[str, str]]]:
-        """Get list of registered schemes for debugging.
-
-        Returns:
-            Dict mapping version to list of {network, scheme} dicts.
-        """
-        result: dict[int, list[dict[str, str]]] = {1: [], 2: []}
-
-        for network, schemes in self._schemes.items():
-            for scheme in schemes:
-                result[2].append({"network": network, "scheme": scheme})
-
-        for network, schemes in self._schemes_v1.items():
-            for scheme in schemes:
-                result[1].append({"network": network, "scheme": scheme})
-
-        return result
 
 
 # ============================================================================
@@ -353,7 +53,7 @@ class _x402ClientBase:
 # ============================================================================
 
 
-class x402Client(_x402ClientBase):
+class x402Client(x402ClientBase):
     """Async client-side component for creating payment payloads.
 
     Supports both sync and async hooks (auto-detected).
@@ -377,12 +77,7 @@ class x402Client(_x402ClientBase):
         self,
         payment_requirements_selector: PaymentRequirementsSelector | None = None,
     ) -> None:
-        """Initialize async x402Client.
-
-        Args:
-            payment_requirements_selector: Custom selector for choosing
-                from filtered requirements. Defaults to first match.
-        """
+        """Initialize async x402Client."""
         super().__init__(payment_requirements_selector)
         # Type the hook lists properly
         self._before_payment_creation_hooks: list[BeforePaymentCreationHook] = []
@@ -398,31 +93,6 @@ class x402Client(_x402ClientBase):
         """Create a new x402Client instance from a configuration object.
 
         Matches TypeScript x402Client.fromConfig() behavior.
-
-        Args:
-            config: The client configuration including schemes, policies,
-                and payment requirements selector.
-
-        Returns:
-            A configured x402Client instance.
-
-        Example:
-            ```python
-            from x402 import x402Client, x402ClientConfig, SchemeRegistration
-            from x402.mechanisms.evm.exact import ExactEvmScheme
-
-            config = x402ClientConfig(
-                schemes=[
-                    SchemeRegistration(
-                        network="eip155:8453",
-                        client=ExactEvmScheme(signer=my_signer),
-                        x402_version=2,
-                    ),
-                ],
-                policies=[prefer_network("eip155:8453")],
-            )
-            client = x402Client.from_config(config)
-            ```
         """
         client = cls(config.payment_requirements_selector)
         for scheme in config.schemes:
@@ -435,50 +105,22 @@ class x402Client(_x402ClientBase):
         return client
 
     # ========================================================================
-    # Hook Registration
+    #
+
     # ========================================================================
 
     def on_before_payment_creation(self, hook: BeforePaymentCreationHook) -> Self:
-        """Register hook to run before payment creation.
-
-        Hook can return AbortResult to abort the operation.
-        Supports both sync and async hooks.
-
-        Args:
-            hook: Hook function (sync or async).
-
-        Returns:
-            Self for chaining.
-        """
+        """Register hook before payment creation. Return AbortResult to abort."""
         self._before_payment_creation_hooks.append(hook)
         return self
 
     def on_after_payment_creation(self, hook: AfterPaymentCreationHook) -> Self:
-        """Register hook to run after successful payment creation.
-
-        Supports both sync and async hooks.
-
-        Args:
-            hook: Hook function (sync or async).
-
-        Returns:
-            Self for chaining.
-        """
+        """Register hook after successful payment creation."""
         self._after_payment_creation_hooks.append(hook)
         return self
 
     def on_payment_creation_failure(self, hook: OnPaymentCreationFailureHook) -> Self:
-        """Register hook to run on payment creation failure.
-
-        Hook can return RecoveredPayloadResult to recover with a payload.
-        Supports both sync and async hooks.
-
-        Args:
-            hook: Hook function (sync or async).
-
-        Returns:
-            Self for chaining.
-        """
+        """Register hook on failure. Return RecoveredPayloadResult to recover."""
         self._on_payment_creation_failure_hooks.append(hook)
         return self
 
@@ -526,130 +168,29 @@ class x402Client(_x402ClientBase):
         resource: ResourceInfo | None,
         extensions: dict[str, Any] | None,
     ) -> PaymentPayload:
-        """Create V2 payment payload."""
-        # 1. Select requirements
-        selected = self._select_requirements_v2(payment_required.accepts)
-
-        # 2. Build context
-        context = PaymentCreationContext(
-            payment_required=payment_required,
-            selected_requirements=selected,
-        )
-
-        # 3. Execute before hooks
-        for hook in self._before_payment_creation_hooks:
-            result = await self._execute_hook(hook, context)
-            if isinstance(result, AbortResult):
-                raise PaymentAbortedError(result.reason)
-
+        """Create V2 payment payload using generator."""
+        gen = self._create_payment_payload_v2_core(payment_required, resource, extensions)
+        result = None
         try:
-            # 4. Find scheme client
-            schemes = find_schemes_by_network(self._schemes, selected.network)
-            if schemes is None or selected.scheme not in schemes:
-                raise SchemeNotFoundError(selected.scheme, selected.network)
-
-            client = schemes[selected.scheme]
-
-            # 5. Create inner payload
-            inner_payload = client.create_payment_payload(selected)
-
-            # 6. Wrap into full PaymentPayload
-            payload = PaymentPayload(
-                x402_version=2,
-                payload=inner_payload,
-                accepted=selected,
-                resource=resource or payment_required.resource,
-                extensions=extensions or payment_required.extensions,
-            )
-
-            # 7. Execute after hooks
-            result_context = PaymentCreatedContext(
-                payment_required=payment_required,
-                selected_requirements=selected,
-                payment_payload=payload,
-            )
-            for hook in self._after_payment_creation_hooks:
-                await self._execute_hook(hook, result_context)
-
-            return payload
-
-        except Exception as e:
-            # Execute failure hooks
-            failure_context = PaymentCreationFailureContext(
-                payment_required=payment_required,
-                selected_requirements=selected,
-                error=e,
-            )
-            for hook in self._on_payment_creation_failure_hooks:
-                result = await self._execute_hook(hook, failure_context)
-                if isinstance(result, RecoveredPayloadResult):
-                    return result.payload  # type: ignore[return-value]
-
-            raise
+            while True:
+                _, hook, ctx = gen.send(result)
+                result = await self._execute_hook(hook, ctx)
+        except StopIteration as e:
+            return e.value
 
     async def _create_payment_payload_v1(
         self,
         payment_required: PaymentRequiredV1,
     ) -> PaymentPayloadV1:
-        """Create V1 payment payload."""
-        # 1. Select requirements
-        selected = self._select_requirements_v1(payment_required.accepts)
-
-        # 2. Build context
-        context = PaymentCreationContext(
-            payment_required=payment_required,
-            selected_requirements=selected,
-        )
-
-        # 3. Execute before hooks
-        for hook in self._before_payment_creation_hooks:
-            result = await self._execute_hook(hook, context)
-            if isinstance(result, AbortResult):
-                raise PaymentAbortedError(result.reason)
-
+        """Create V1 payment payload using generator."""
+        gen = self._create_payment_payload_v1_core(payment_required)
+        result = None
         try:
-            # 4. Find scheme client
-            schemes = find_schemes_by_network(self._schemes_v1, selected.network)
-            if schemes is None or selected.scheme not in schemes:
-                raise SchemeNotFoundError(selected.scheme, selected.network)
-
-            client = schemes[selected.scheme]
-
-            # 5. Create inner payload
-            inner_payload = client.create_payment_payload(selected)
-
-            # 6. Wrap into full PaymentPayloadV1
-            payload = PaymentPayloadV1(
-                x402_version=1,
-                scheme=selected.scheme,
-                network=selected.network,
-                payload=inner_payload,
-            )
-
-            # 7. Execute after hooks
-            result_context = PaymentCreatedContext(
-                payment_required=payment_required,
-                selected_requirements=selected,
-                payment_payload=payload,
-            )
-            for hook in self._after_payment_creation_hooks:
-                await self._execute_hook(hook, result_context)
-
-            return payload
-
-        except Exception as e:
-            # Execute failure hooks
-            failure_context = PaymentCreationFailureContext(
-                payment_required=payment_required,
-                selected_requirements=selected,
-                error=e,
-            )
-            for hook in self._on_payment_creation_failure_hooks:
-                result = await self._execute_hook(hook, failure_context)
-                if isinstance(result, RecoveredPayloadResult):
-                    return result.payload  # type: ignore[return-value]
-
-            raise
+            while True:
+                _, hook, ctx = gen.send(result)
+                result = await self._execute_hook(hook, ctx)
+        except StopIteration as e:
+            return e.value
 
     async def _execute_hook(self, hook: Any, context: Any) -> Any:
         """Execute hook, auto-detecting sync/async."""
@@ -664,7 +205,7 @@ class x402Client(_x402ClientBase):
 # ============================================================================
 
 
-class x402ClientSync(_x402ClientBase):
+class x402ClientSync(x402ClientBase):
     """Sync client-side component for creating payment payloads.
 
     Only supports sync hooks. For async hook support, use x402Client.
@@ -686,12 +227,7 @@ class x402ClientSync(_x402ClientBase):
         self,
         payment_requirements_selector: PaymentRequirementsSelector | None = None,
     ) -> None:
-        """Initialize sync x402Client.
-
-        Args:
-            payment_requirements_selector: Custom selector for choosing
-                from filtered requirements. Defaults to first match.
-        """
+        """Initialize sync x402Client."""
         super().__init__(payment_requirements_selector)
         # Type the hook lists for sync-only
         self._before_payment_creation_hooks: list[SyncBeforePaymentCreationHook] = []
@@ -707,31 +243,6 @@ class x402ClientSync(_x402ClientBase):
         """Create a new x402ClientSync instance from a configuration object.
 
         Matches TypeScript x402Client.fromConfig() behavior (sync variant).
-
-        Args:
-            config: The client configuration including schemes, policies,
-                and payment requirements selector.
-
-        Returns:
-            A configured x402ClientSync instance.
-
-        Example:
-            ```python
-            from x402 import x402ClientSync, x402ClientConfig, SchemeRegistration
-            from x402.mechanisms.evm.exact import ExactEvmScheme
-
-            config = x402ClientConfig(
-                schemes=[
-                    SchemeRegistration(
-                        network="eip155:8453",
-                        client=ExactEvmScheme(signer=my_signer),
-                        x402_version=2,
-                    ),
-                ],
-                policies=[prefer_network("eip155:8453")],
-            )
-            client = x402ClientSync.from_config(config)
-            ```
         """
         client = cls(config.payment_requirements_selector)
         for scheme in config.schemes:
@@ -748,46 +259,17 @@ class x402ClientSync(_x402ClientBase):
     # ========================================================================
 
     def on_before_payment_creation(self, hook: SyncBeforePaymentCreationHook) -> Self:
-        """Register hook to run before payment creation.
-
-        Hook can return AbortResult to abort the operation.
-        Note: Only sync hooks are supported. Use x402Client for async hooks.
-
-        Args:
-            hook: Sync hook function.
-
-        Returns:
-            Self for chaining.
-        """
+        """Register hook before payment creation. Return AbortResult to abort."""
         self._before_payment_creation_hooks.append(hook)
         return self
 
     def on_after_payment_creation(self, hook: SyncAfterPaymentCreationHook) -> Self:
-        """Register hook to run after successful payment creation.
-
-        Note: Only sync hooks are supported. Use x402Client for async hooks.
-
-        Args:
-            hook: Sync hook function.
-
-        Returns:
-            Self for chaining.
-        """
+        """Register hook after successful payment creation."""
         self._after_payment_creation_hooks.append(hook)
         return self
 
     def on_payment_creation_failure(self, hook: SyncOnPaymentCreationFailureHook) -> Self:
-        """Register hook to run on payment creation failure.
-
-        Hook can return RecoveredPayloadResult to recover with a payload.
-        Note: Only sync hooks are supported. Use x402Client for async hooks.
-
-        Args:
-            hook: Sync hook function.
-
-        Returns:
-            Self for chaining.
-        """
+        """Register hook on failure. Return RecoveredPayloadResult to recover."""
         self._on_payment_creation_failure_hooks.append(hook)
         return self
 
@@ -835,130 +317,29 @@ class x402ClientSync(_x402ClientBase):
         resource: ResourceInfo | None,
         extensions: dict[str, Any] | None,
     ) -> PaymentPayload:
-        """Create V2 payment payload."""
-        # 1. Select requirements
-        selected = self._select_requirements_v2(payment_required.accepts)
-
-        # 2. Build context
-        context = PaymentCreationContext(
-            payment_required=payment_required,
-            selected_requirements=selected,
-        )
-
-        # 3. Execute before hooks
-        for hook in self._before_payment_creation_hooks:
-            result = self._execute_hook_sync(hook, context)
-            if isinstance(result, AbortResult):
-                raise PaymentAbortedError(result.reason)
-
+        """Create V2 payment payload using generator."""
+        gen = self._create_payment_payload_v2_core(payment_required, resource, extensions)
+        result = None
         try:
-            # 4. Find scheme client
-            schemes = find_schemes_by_network(self._schemes, selected.network)
-            if schemes is None or selected.scheme not in schemes:
-                raise SchemeNotFoundError(selected.scheme, selected.network)
-
-            client = schemes[selected.scheme]
-
-            # 5. Create inner payload
-            inner_payload = client.create_payment_payload(selected)
-
-            # 6. Wrap into full PaymentPayload
-            payload = PaymentPayload(
-                x402_version=2,
-                payload=inner_payload,
-                accepted=selected,
-                resource=resource or payment_required.resource,
-                extensions=extensions or payment_required.extensions,
-            )
-
-            # 7. Execute after hooks
-            result_context = PaymentCreatedContext(
-                payment_required=payment_required,
-                selected_requirements=selected,
-                payment_payload=payload,
-            )
-            for hook in self._after_payment_creation_hooks:
-                self._execute_hook_sync(hook, result_context)
-
-            return payload
-
-        except Exception as e:
-            # Execute failure hooks
-            failure_context = PaymentCreationFailureContext(
-                payment_required=payment_required,
-                selected_requirements=selected,
-                error=e,
-            )
-            for hook in self._on_payment_creation_failure_hooks:
-                result = self._execute_hook_sync(hook, failure_context)
-                if isinstance(result, RecoveredPayloadResult):
-                    return result.payload  # type: ignore[return-value]
-
-            raise
+            while True:
+                _, hook, ctx = gen.send(result)
+                result = self._execute_hook_sync(hook, ctx)
+        except StopIteration as e:
+            return e.value
 
     def _create_payment_payload_v1(
         self,
         payment_required: PaymentRequiredV1,
     ) -> PaymentPayloadV1:
-        """Create V1 payment payload."""
-        # 1. Select requirements
-        selected = self._select_requirements_v1(payment_required.accepts)
-
-        # 2. Build context
-        context = PaymentCreationContext(
-            payment_required=payment_required,
-            selected_requirements=selected,
-        )
-
-        # 3. Execute before hooks
-        for hook in self._before_payment_creation_hooks:
-            result = self._execute_hook_sync(hook, context)
-            if isinstance(result, AbortResult):
-                raise PaymentAbortedError(result.reason)
-
+        """Create V1 payment payload using generator."""
+        gen = self._create_payment_payload_v1_core(payment_required)
+        result = None
         try:
-            # 4. Find scheme client
-            schemes = find_schemes_by_network(self._schemes_v1, selected.network)
-            if schemes is None or selected.scheme not in schemes:
-                raise SchemeNotFoundError(selected.scheme, selected.network)
-
-            client = schemes[selected.scheme]
-
-            # 5. Create inner payload
-            inner_payload = client.create_payment_payload(selected)
-
-            # 6. Wrap into full PaymentPayloadV1
-            payload = PaymentPayloadV1(
-                x402_version=1,
-                scheme=selected.scheme,
-                network=selected.network,
-                payload=inner_payload,
-            )
-
-            # 7. Execute after hooks
-            result_context = PaymentCreatedContext(
-                payment_required=payment_required,
-                selected_requirements=selected,
-                payment_payload=payload,
-            )
-            for hook in self._after_payment_creation_hooks:
-                self._execute_hook_sync(hook, result_context)
-
-            return payload
-
-        except Exception as e:
-            # Execute failure hooks
-            failure_context = PaymentCreationFailureContext(
-                payment_required=payment_required,
-                selected_requirements=selected,
-                error=e,
-            )
-            for hook in self._on_payment_creation_failure_hooks:
-                result = self._execute_hook_sync(hook, failure_context)
-                if isinstance(result, RecoveredPayloadResult):
-                    return result.payload  # type: ignore[return-value]
-
-            raise
+            while True:
+                _, hook, ctx = gen.send(result)
+                result = self._execute_hook_sync(hook, ctx)
+        except StopIteration as e:
+            return e.value
 
     def _execute_hook_sync(self, hook: Any, context: Any) -> Any:
         """Execute hook synchronously. Raises if async hook detected."""
